feat: add create normalized matrices

+++ src/database/entities/reccomendation.entity.ts
@@ -36,7 +36,7 @@ export class ReccomendationEntity
     scale: 2,
     transformer: {
       to: (value: number) => value.toFixed(2),
-      from: (value: string) => parseInt(value),
+      from: (value: string) => parseFloat(value),
     },
   })
   reccomendationScore: number;
